create separate api to create track
